Feat: Started Flask setup & web integration.
- Started the process of setting up and integrating Flask.
- Started migration of DB from local to web.
- Redesigned project file structure to accommodate Flask integration and DB migration.
