+++ expense_tracker/models/base.py
@@ -1,0 +1,4 @@
+from sqlalchemy.ext.declarative import declarative_base
+
+
+Base = declarative_base()
