+++ requirements.txt
@@ -12,3 +12,4 @@ alembic~=1.13.2
 pytest~=8.3.1
 pytest-mock==3.14.0
 cryptography~=43.0.0
+Flask~=3.0.3
